# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mobile Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SALES / INVOICES ─────────────────────────────────────────
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleDetail"
    delete:
      operationId: deleteSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── RETURNS ─────────────────────────────────────────────────
...
  /software-services:
    get:
      operationId: listSoftwareServices
      tags: [software]
      responses:
        "200":
          description: Software services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SoftwareService"
    post:
      operationId: createSoftwareService
      tags: [software]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareService"
  /software-services/{id}:
    patch:
      operationId: updateSoftwareService
      tags: [software]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareService"

  # ─── EXPENSES ────────────────────────────────────────────────
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      responses:
        "200":
          description: Wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
    post:
      operationId: createWalletTransaction
      tags: [wallets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransaction"

  # ─── BANK ACCOUNTS ───────────────────────────────────────────
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        totalPurchases:
          type: number
        totalDebt:
          type: number
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string

    # ─── SALES ──────────────────────────────────────────────────
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, discount, total]
      properties:
        productId:
          type: ["integer", "null"]
        deviceId:
          type: ["integer", "null"]
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
    SaleItemInput:
      type: object
      required: [productName, quantity, unitPrice, discount, total]
      properties:
        productId:
          type: integer
        deviceId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
    Sale:
      type: object
      required: [id, invoiceNumber, customerId, customerName, subtotal, discount, total, paymentMethod, paidAmount, dueAmount, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, etisalat_cash, we_pay, instapay, bank_transfer]
        paidAmount:
          type: number
        dueAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SaleDetail:
      type: object
      required: [id, invoiceNumber, customerId, customerName, subtotal, discount, total, paymentMethod, paidAmount, dueAmount, items, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        paidAmount:
          type: number
        dueAmount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [customerName, items, paymentMethod, paidAmount]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, etisalat_cash, we_pay, instapay, bank_transfer]
        paidAmount:
          type: number
        notes:
          type: string

    # ─── RETURNS ─────────────────────────────────────────────────
...
    SoftwareService:
      type: object
      required: [id, serviceType, customerName, customerPhone, cost, salePrice, status, createdAt]
      properties:
        id:
          type: integer
        serviceType:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        deviceBrand:
          type: ["string", "null"]
        deviceModel:
          type: ["string", "null"]
        cost:
          type: number
        salePrice:
          type: number
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, done]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SoftwareServiceInput:
      type: object
      required: [serviceType, customerName, customerPhone, cost, salePrice]
      properties:
        serviceType:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deviceBrand:
          type: string
        deviceModel:
          type: string
        cost:
          type: number
        salePrice:
          type: number
        technicianId:
          type: integer
        notes:
          type: string
    SoftwareServiceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, done]
        cost:
          type: number
        salePrice:
          type: number
        technicianId:
          type: integer
        notes:
          type: string

    # ─── EXPENSES ────────────────────────────────────────────────
...
    Wallet:
      type: object
      required: [id, name, company, phoneNumber, balance]
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
          enum: [vodafone, etisalat, we, instapay, other]
        phoneNumber:
          type: string
        balance:
          type: number
    WalletInput:
      type: object
      required: [name, company, phoneNumber]
      properties:
        name:
          type: string
        company:
          type: string
          enum: [vodafone, etisalat, we, instapay, other]
        phoneNumber:
          type: string
        balance:
          type: number
    WalletTransaction:
      type: object
      required: [id, walletId, type, amount, balanceBefore, balanceAfter, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        type:
          type: string
          enum: [deposit, withdraw, transfer, receive]
        amount:
          type: number
        fee:
          type: number
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    WalletTransactionInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [deposit, withdraw, transfer, receive]
        amount:
          type: number
        fee:
          type: number
        notes:
          type: string

    # ─── BANK ACCOUNTS ───────────────────────────────────────────
...